* tta/swig/perl/_Texinfo_Tests.pm: do not create directory and link if the .libs directory does not exist, assuming that the MakeMaker build system is used in that case.
* tta/swig/perl/Makefile.PL (INC): update compile flags.

+++ tta/swig/perl/_Texinfo_Tests.pm
@@ -42,18 +42,20 @@ if (!defined($t2a_builddir) and defined($srcdir)) {
 }
 if (defined($t2a_builddir)) {
   my $xsdir = join('/', ($t2a_builddir, 'swig', 'perl', '.libs'));
-  unshift @INC, $xsdir;
-  # XSLoader searches in auto/Texinfo, so make a symlink from
-  # auto/Texinfo to ../ to get back to .libs
-  my $autodir = join('/', ($xsdir, 'auto'));
-  if (!-d $autodir) {
-    mkdir ($autodir) or die "Failed to mkdir $autodir: $!\n";;
-  }
-  my $loaddir = join('/', ($autodir, 'Texinfo'));
-  if (-e $loaddir) {
-    unlink($loaddir) or die "Failed to remove file $loaddir: $!\n";
-  }
-  symlink($updir, $loaddir);
+  if (-d $xsdir) {
+    unshift @INC, $xsdir;
+    # XSLoader searches in auto/Texinfo, so make a symlink from
+    # auto/Texinfo to ../ to get back to .libs
+    my $autodir = join('/', ($xsdir, 'auto'));
+    if (!-d $autodir) {
+      mkdir ($autodir) or die "Failed to mkdir $autodir: $!\n";;
+    }
+    my $loaddir = join('/', ($autodir, 'Texinfo'));
+    if (-e $loaddir) {
+      unlink($loaddir) or die "Failed to remove file $loaddir: $!\n";
+    }
+    symlink($updir, $loaddir);
+  } # if the directory is not present, we assume a MakeMaker build
 }
 
 eval { require Text::Diff; Text::Diff->import('diff'); };
